Split device protection into per-OS modules and start the Linux wallet
